Put most of view implementation in common
